# reference-implementation/openapi/reference-full.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "PDPP Reference Implementation (full, includes /_ref)",
    "version": "0.1.0",
    "description": "Public PDPP JSON APIs plus reference-designated /_ref operator/control surfaces."
  },
...
  "paths": {
...
    "/grants/{grantId}/revoke": {
      "post": {
        "operationId": "revokeGrant",
        "tags": [
          "grants"
        ],
        "summary": "Revoke a grant and all tokens minted from it.",
        "parameters": [
          {
            "in": "path",
            "name": "grantId",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false,
                  "properties": {
                    "revoked": {
                      "const": true
                    }
                  },
                  "required": [
                    "revoked"
                  ]
                }
              }
            }
          },
          "403": {
            "description": "Grant is malformed or no longer valid",
            "content": {
              "application/json": {
                "schema": {
                  "$id": "pdpp/common/PdppError",
                  "type": "object",
                  "additionalProperties": false,
                  "properties": {
                    "error": {
                      "type": "object",
                      "additionalProperties": false,
                      "properties": {
                        "type": {
                          "type": "string"
                        },
                        "code": {
                          "type": "string"
                        },
                        "message": {
                          "type": "string"
                        },
                        "param": {
                          "type": "string"
                        },
                        "request_id": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "type",
                        "code",
                        "message",
                        "request_id"
                      ]
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          }
        }
      }
    },
...
    "/_ref/connectors/{connectorId}": {
      "get": {
        "operationId": "refGetConnector",
        "tags": [
          "reference",
          "connectors"
        ],
        "summary": "Get a single connector with manifest excerpt, schedule, recent runs, and stream summaries.",
        "parameters": [
          {
            "in": "path",
            "name": "connectorId",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "properties": {
                    "connector_id": {
                      "type": "string"
                    },
                    "display_name": {
                      "type": "string"
                    },
                    "manifest_version": {
                      "type": "string"
                    },
                    "streams": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    },
                    "freshness": {
                      "$id": "pdpp/common/Freshness",
                      "type": "object",
                      "additionalProperties": false,
                      "properties": {
                        "status": {
                          "$id": "pdpp/common/FreshnessStatus",
                          "type": "string",
                          "enum": [
                            "current",
                            "stale",
                            "unknown"
                          ]
                        },
                        "captured_at": {
                          "type": "string",
                          "format": "date-time"
                        },
                        "last_attempted_at": {
                          "type": "string",
                          "format": "date-time"
                        }
                      },
                      "required": [
                        "status"
                      ]
                    },
                    "schedule": {
                      "type": [
                        "object",
                        "null"
                      ],
                      "additionalProperties": true,
                      "properties": {
                        "interval_seconds": {
                          "type": "integer"
                        },
                        "jitter_seconds": {
                          "type": "integer"
                        },
                        "enabled": {
                          "type": "boolean"
                        },
                        "next_due_at": {
                          "type": [
                            "string",
                            "null"
                          ]
                        }
                      }
                    },
                    "last_run": {
                      "type": [
                        "object",
                        "null"
                      ],
                      "additionalProperties": true,
                      "properties": {
                        "run_id": {
                          "type": "string"
                        },
                        "status": {
                          "type": "string"
                        },
                        "started_at": {
                          "type": "string"
                        },
                        "finished_at": {
                          "type": [
                            "string",
                            "null"
                          ]
                        }
                      }
                    }
                  },
                  "required": [
                    "connector_id"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$id": "pdpp/common/PdppError",
                  "type": "object",
                  "additionalProperties": false,
                  "properties": {
                    "error": {
                      "type": "object",
                      "additionalProperties": false,
                      "properties": {
                        "type": {
                          "type": "string"
                        },
                        "code": {
                          "type": "string"
                        },
                        "message": {
                          "type": "string"
                        },
                        "param": {
                          "type": "string"
                        },
                        "request_id": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "type",
                        "code",
                        "message",
                        "request_id"
                      ]
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$id": "pdpp/common/PdppError",
                  "type": "object",
                  "additionalProperties": false,
                  "properties": {
                    "error": {
                      "type": "object",
                      "additionalProperties": false,
                      "properties": {
                        "type": {
                          "type": "string"
                        },
                        "code": {
                          "type": "string"
                        },
                        "message": {
                          "type": "string"
                        },
                        "param": {
                          "type": "string"
                        },
                        "request_id": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "type",
                        "code",
                        "message",
                        "request_id"
                      ]
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          },
          "409": {
            "description": "Conflict (e.g. run_already_active)",
            "content": {
              "application/json": {
                "schema": {
                  "$id": "pdpp/common/PdppError",
                  "type": "object",
                  "additionalProperties": false,
                  "properties": {
                    "error": {
                      "type": "object",
                      "additionalProperties": false,
                      "properties": {
                        "type": {
                          "type": "string"
                        },
                        "code": {
                          "type": "string"
                        },
                        "message": {
                          "type": "string"
                        },
                        "param": {
                          "type": "string"
                        },
                        "request_id": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "type",
                        "code",
                        "message",
                        "request_id"
                      ]
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}